Class cannot be fixtures
